pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.9.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.3",
]
